requirements.txt
PyYAML==6.0.1
watchdog==4.0.0
pytest==7.4.3
